+++ src/jOpqua.jl
@@ -8,6 +8,7 @@ include("simulation/structs.jl")
 include("parameters/parameters.jl")
 
 include("simulation/initializers.jl")
+include("simulation/choice.jl")
 include("simulation/weights.jl")
 include("simulation/events.jl")
 include("simulation/simulation.jl")